2 new units

+++ Assets/Scripts/AI/Move/Common.cs
@@ -14,7 +14,7 @@ namespace AIPlayer
             public static T RandomOptionOfBest<T>(List<T> options) where T : Option
             {
                 options = options.OrderByDescending(op => op.score).ToList();
-                int highest = options[0].score;
+                float highest = options[0].score;
 
                 List<T> possible = new List<T>();
                 foreach (T op in options)
